Display a Message then Read a Line from the Terminal

Display a Message then Read a Line from the Terminal

Usage

cli_menu(
  choices,
  title = NULL,
  message = "Enter your selection or press {.kbd 0} to exit.",
  prompt = "Selection:",
  labels = "arabic",
  sep = ": ",
  bullet = "",
  exit = "0",
  ind = FALSE,
  call = .envir,
  .envir = parent.frame()
)

Arguments

choices

Required list of choices. If named, use the names as the choice options.

title

Title for menu. Passed to cli::cli_h2()

message

Additional message to display after choices and before prompt.

prompt

Characters to show as user prompt in console following displayed menu options. Defaults to "Selection:" (matching utils::menu())

labels

Style to use for menu choices (from isstatic::as_numbered_labels). Defaults to "arabic". Ignored if choices are a named vector or list.

sep

Separator between choice name and value. Ignored if choices are a named vector or list.

bullet

Character to use for bullet. Defaults to "".

exit

Character to use to exit menu.

call

The execution environment of a currently running function, e.g. call = caller_env(). The corresponding function call is retrieved and mentioned in error messages as the source of the error.

You only need to supply call when throwing a condition from a helper function which wouldn't be relevant to mention in the message.

Can also be NULL or a defused function call to respectively not display any call or hard-code a code to display.

For more information about error calls, see Including function calls in error messages.

.envir

Environment to evaluate the glue expressions in.

Examples

if (FALSE) {
if (interactive()) {
  cli_menu(list("A", "B", "C"), title = "Pick a letter?")

  cli_menu(c("A", "B", "C"), labels = "Alpha")

  cli_menu(list(list(1, 2, 3), "A", "B", "C"), title = "Pick from a list")
}
}